requirements.txt
streamlit
requests
pandas
beautifulsoup4
python-docx
